WL#3776 : Consistent snapshot backup for MVCC
This patch adds consistent snapshot algorithm for online backup. The
consistent snapshot algorithm is designed to work with any storage engine
that supports a consistent read operation (transaction). Currently, this
is limited to InnoDB but the code was written to allow the use of any
storage engine that supposts consistent read via the consistent shapshot
method in the handler.