Bug#27249 table_wild with alias: select t1.* as something
Aliases to table wildcards are silently ignored, but they should
not be allowed as it is non-standard and currently useless. There
is not point in having a alias to a wildcard of column names.

The solution is to rewrite the select_item rule so that aliases
for table wildcards are not accepted.

Contribution by Martin Friebe