WL#4060 (kernel updates for beta release):
This patch modifies backup kernel code so that it uses the backup stream library 
for writing and reading backup images. Now backup kernel writes backup images in
the format described in WL#4063.
  
Limitations:
 
1. Errors are detected but not always reported in the error log - only a general 
error message will be shown.
  
2. Object names using non-standard charsets are still not supported. The code 
doesn't address this issue yet.
  
3. The default server charset is stored in the image, but it is not set upon 
restore. This should not be a big problem however, since I believe that CREATE 
TABLE statements we use always set charset and collaction explicitly. 
  
4. I don't obtain CREATE DATABASE statements from the server and thus don't save 
things like default charset or collation. It should be easy to add but I simply 
didn't have time to do it now. There is a bug entry for this.
  
5. The summary produced by BACKUP/RESTORE commands is now very minimalistic. Due 
to complex buffering in the backup stream library it is difficult to get exact 
sizes of various parts of the backup image and thus produce similar summary as 
before. Since format of the summary is going to change soon, I didn't spend time 
on improving it.
  
6. Eventually, the endianess of the server which created backup image should be 
stored in image flags. This code doesn't store this information (as I'm not sure 
how to get it). Note however that the image format *is* endianess agnostic (at 
least the part written by the kernel). The purpose of the flag is to provide 
more information for the user and possibly for restore drivers (not supported in 
the current driver API).