Bug#30480: Falcon: searches fail if LIKE and key partition (also fixes the bugs: 29320, 29493 and 30536)
Problem: Partitioning did not handle unordered scans correctly
for engines with unordered read order.

Solution: do not stop scanning if a record is out of range, since
there can be more records within the range afterwards.

Note. this is the test case ONLY, the real patch is separate since it
goes into 5.1 as well (for non mysql storage engines,
since falcon is in 6.0)