BUG#33055 (Replication fails for UPDATE when using falcon Storage engine):
The read- and write-sets were not respected when deciding what fields to
transfer to the slave, nor when writing the row to the handler on the sla-
ve side.

This patch adds code that only writes the columns actually needed to be 
able to replicate row-based. This means that the columns in the write set
of the table is used for ha_write_row(), the columns in the read set for
ha_delete_row(), and for ha_update_row() the columns in the read set is
used for the before image, and the columns of the write set for the after
image.

In addition, for cursor-based delete and update, the necessary bits are
set to be able to get enough information to locate the affected row on
the slave side, which means that it behaves as if HA_PRIMARY_KEY_REQUIRED_
FOR_DELETE was set for the table when row-based logging is active.