Bug#33480 - Two kinds of hangs / deadlocks are fixed here.
1) In ha_partition.cc, prevent serialization into Falcon of 
threads writing to autoincrement fields.  Only InnoDB needs 
this.  It can cause deadlocks and timeouts in Falcon which 
appear as 0% CPU utilization when all threads are hung.  
When this happens for just a few threads, the serial log 
can grow indefinitely.
  
2) If a deadlock happens between threads which are checking 
for duplicates, the code in Table::checkUniqueRecordVersion()
was not handling the detected deadlock by throwing an exception.  
Now it does.  This can cause 100% CPU utilization hangs between 
threads.  One is waiting and the other is looping in 
Table::checkUnigueIndexes().  This can also cause the serial log
to grow indefinitely.