Bug#33041 : cannot set FALCON_CONSISTENT_READ for local sessions.
Solution is to change the scope of the variable to be connection-specific
such that SET FALCON_CONSISTENT_READ=<ON|OFF> modifies the setting
only for current connection.

SET GLOBAL FALCON_CONSISTENT_READ still works, but it affects only 
current and new connections. Currently it is not possible to disable 
"SET GLOBAL" ,as plugin API does not provide possibility to define 
local-only scope.