Bug#31765 (BACKUP DATABASE broken syntax)
Removed un necessary reduces in the middle of the 'backup' rule.
Removed un necessary reduces in the middle of the 'restore' rule.
Simplifying the grammar improves the code, and performances by making the
parser state automaton smaller.

Fixed the 'database_list' rule to:
- either expand to a '*',
- or expand to a list of database names.
but not mix both, which caused the reported bug.