BUG#34858 Server crashes by backup tests
Patch fixes problems encountered when merging runtime code changes
into main for online backup.

The problem encountered caused online backup to crash the server. 
The cause originated from two locations. The snapshot driver was not
closing the transaction it uses for the consistent read and the 
backup progress tables were forcing an assertion on table->in_use.
The assertion was included in the server code to guard against the
very thing that the backup progress code was doing: open a table in
one thread and write to it in another. This has been corrected in 
this patch.

Changes made include:

* snapshot driver fixed
* progress table access changed to that of performance tables (logs)