BUG#35548: NDB broken multi_range_read implementation in 6.0
For MRR unordered scans, rows are returned in random order, and custom
pointer returned to upper layer was not correctly associated with the
range that each row originated from.

Fixed by storing custom pointers at the start of the MRR buffer in a
fixed-format array allowing random access to the correct pointer.