Bug#34890 - hold the SyncObject::mutex to protect thread->lockGranted and send it into thread->sleep so that it will overlap the PTHREADS mutex.
This prevents a race condition where a thread is woken up inadvertently,
and while it is looping around to go back to sleep, it somehow misses
the event or signal by the thread holding the lock.  When this happens,
the threads waits on that SyncObject until another thread locks and 
releases it.  This is a thread stall.