Bug#34778, "Possible memory leak running UPDATEs in tight loop"
This test generates thousands of prior record versions for just a
very few records. The record scavenger ignored the prior record chain
because the base record was not a scavenge candidate.

The scavenger now scavenges prior records that match the transaction and
record generation criteria, even when the base record is not scavengable.