BUG#33571 (Backup: restore failure if sql_mode=ansi)
This patch modifies object services methods so that sql_mode is locally changed to
default when operations which depend on it are executed. These operations are 
obtaining the serialization string (format of SHOW CREATE output depends on sql_mode)
and "executing" an object (syntax of CREATE statement depends on sql_mode).