Additional fix for patch that has changed type of Key_part_spec::field_name and Key::name to LEX_STRING. (The latter was a prerequisite patch for the 2nd milestone of WL#148 "Foreign keys").
In the main patch I've missed one of the places which should have been
updated to use Key::name.str instead of Key::name. As result we were
getting garbage as key name in one of error messages in create.test
on Win64 and Sparc. This additional change fixes this omission.