BUG#33723 - ALTER TABLE into non-existing Falcon tablespace blocks             further ALTERs BUG#34048 - Falcon: after error with tablespace, I can't create table BUG#34617 - Falcon assertion in StorageHandler::addTable, line 622
If Falcon fails to create a table for some reason (e.g. non-existent
tablespace), it is not possible to create a table in the same database
with the same name anymore (until server is restarted). Or assertion
failure (=server crash) may occur for further create table statements.

The above is true for ALTER TABLE as well, when it fails to create
temporary table.

The problem was that table share was not removed from table share
hash on error.