BUG#34721 (Backup driver can't refuse to provide driver):
The problem was that a storage engine could not refuse to create a native backup 
engine without stopping the whole backup process with an error. The intended 
behaviour is to use a built-in backup engine for that storage engine is such 
case.

This patch modifies the backup kernel so that it falls back to the
built-in backup engines in case a native backup engine can not be created.

In detail each storage engine defines a get_backup_engine() factory
function (inside the handlerton). If that function fails to create a
backup engine then the backup engine selection logic does now push a
warning and continues using a built-in backup engine for this storage
engine.