Bug#32140: wrong error code caught when an SF() call is interruped with KILL query
The problem is that killing a query which calls a stored function
could return a wrong error (table corrupt) instead of the query
interrupted error message.

The solution is to not set the table corrupt error if the query
is killed, the query interrupted error message will be set  later
when the query is finished.