WL#4347 (backup kernel extension for handling tablespaces):
This patch extends the current kernel so that it handles tables which use 
tablespaces as specified in the WL. This patch assumes that required extensions 
of server object services and backup stream library are already in the tree.

The main changes are:

- Extending Image_info class so that tablespaces can be stored inside the 
  catalogue.
- Define/extend backup catalogue iterators to include tablespaces.
- Take care of tables which use tablespaces when adding them to backup 
  catalogue.
- Re-create tablespaces during restore.

The patch also adds error reporting in places where it was missing.