Bug#35936 (Garbage in syntax error message)
Before this fix, the default initialization of
Lex_input_stream::lookahead_token was using the value END_OF_INPUT
to represent that no lookahead token has been parsed.

This choice was unfortunate, since END_OF_INPUT is a real token value
that can be returned by the lexer, which caused confusion when a query contains
the following tokens:
  <WITH> <END_OF_INPUT>
which in turn happens when a (malformed) query ends with <WITH>.

Changed the value that represents "no token" to an out-of-bound token value,
(-1), that can not be returned by the lexer.