Fix for bug #33969: Updating a text field via a left join
When creating a temporary TEXT/BLOB field from an Item in
Item::make_string_field(), the field's type was unconditionally set
to the one corresponding to the maximum length (i.e. LONGTEXT/
LONGBLOB). This resulted in problems when exactly the same TEXT/BLOB
is type required in cases like CREATE ... SELECT or creating internal
temporary tables for joins.

Fixed by calling a different constructor for Field_blob so that an
appropriate type is used depending on the Item's max_length value.