WL#4259 - Debug Sync Facility
The Debug Sync Facility allows to place synchronization points
in the code:

open_tables(...)

DEBUG_SYNC(thd, "after_open_tables");

lock_tables(...)

When activated, a sync point can

- Send a signal and/or
- Wait for a signal

Nomenclature:

- signal:            A value of a global variable that persists
                     until overwritten by a new signal. The global
                     variable can also be seen as a "signal post"
                     or "flag mast". Then the signal is what is
                     attached to the "signal post" or "flag mast".

- send a signal:     Assign the value (the signal) to the global
                     variable ("set a flag") and broadcast a
                     global condition to wake those waiting for
                     a signal.

- wait for a signal: Loop over waiting for the global condition until
                     the global value matches the wait-for signal.

Please find more information in the top comment in debug_sync.cc
or in the worklog entry.

Included are modified MERGE table tests. They can be seen as
examples for the use of the facility. The merge-big test case is
removed. The contained test is moved to the new merge-sync test.