A fix for Bug#35395: Object services for enumerating view dependencies make server deadlock.
The problem was that ViewBaseObjectsIterator::create() method opened
a view and its underlying tables in order to get a list of all
the tables on which view depends without closing them afterwards.
This happened due to false assumption that THD::~THD() closes all
the tables which were open for THD object being destroyed.

The fix is to call close_thread_tables() explicitly. Also, since
we're interested only in meta-data, there is no need to lock tables.
So, open_tables() can be used instead of open_and_lock_tables().

A test case is not provided because there is no easy way to test
this without changing the source code.