Bug #35546 mysqlbinlog.cc: Load_log_processor::process_first_event() fails creating unique Bug #35543 mysqlbinlog.cc does not properly work with tmp files
There were failures on pb executions in that mysqlbinlog could not compose a new 
unique file name of its purpose.
The reason of the bug has been that mysqlbinlog used the system wide temprorary directory instead
of the specified by mtr so that mysqlbinlog's temporary files happened to be out of reach of mtr.

Logics of the mysqlbinlog program correctly allows leaving new produced files such as ones that 
hold LOAD DATA data.
This is fine as long as a caller of mysqlbinlog takes care of the new files later.

The problem fixed with specifying explicitly the temp directory for mysqlbinlog when it's called
from mtr. New files mysqlbinlog can produce now reside in the mtr temporary directory.

Also there were memory leaks reported (bug#35543) which are fixed with finding out the sources and deploying
freeing.