BUG#33355 : Backup: hang if database is mysql
Corrects issue found concerning case sensitive platforms. On case 
sensitive platforms it was not possible to backup or restore a
database with the name 'MYSQL' (or any other case variant) because the
code was doing a case insensitive compare.

This patch allows the use of case variant names like 'MYSQL' as a
database for systems that have case sensitive identifiers.