Fixed bug #35685.
This bug in the function JOIN_CACHE_BKA::init could manifest itself
when key expressions for a ref access depended on columns not only
from the previous join table. The bug usually caused an assertion
abort for the debug version.

Some cosmetic change to the function make_join_readinfo was added.