WL#4211 (Online Backup: Dependency checking)
The patch implements design described in the worklog. In short, all objects added to
the backup catalogue are also put on a dependency list ensuring correct ordering 
among them. This list determines the order in which object metadata is saved in the 
backup image and thus the order in which they are restored.