WL#3726 "DDL locking for all metadata objects".
Fix for sporadical flush.test failures due to failing assert.

The fact that in open_ltable() we forgot to release metadata
lock before calling handle_failed_open_table_attempt() when
open_table() encountered old version of share and requested
backing off led to assertion failure in mdl_wait_for_locks(),
which assumes that lock requests for which we perform wait are
not satisfied.

This fix ensures that we release metadata lock in such case.