A fix for Bug#36703: Unknown symbol EncodedDataStream::decode when linking mysqld with Sun Studio
The problem was due to Sun Studio compiler produced a different symbol names in the object files for the decode()
method for the implementation of the method and when it was called.
This was triggered due to that the declartion of the method used "unsigned char" while the definition of the method used a UCHAR (which is typedef unsigned char). 

Fixed this problem by using UCHAR in both declaration and definition.