Fix to the second of the memory alignment issues in Bug#36745 falcon crash on solaris
The problem that caused crashes on SPARC running in 64 bit mode was due a memory aligment issue
when "nodes" were allocated in the DeferredIndex::initialSpace. Memory is allocated in the
initialSpace starting from the end of the memory area. Since the end of the initialSpace was not
aligned on an address boundary this resulted in that memory allocated from it was not aligned and thus
crashed on Solaris when running in 64 bit mot on SPARC. Fixed this by ensuring that the size
of the initialSpace is "8 byte memory aligned" (increased the size of it from 500 bytes to 512 bytes).