Bug#36396 - Assertion in IO::pread (on closed tablespace file) The problem was that gopher  processing "drop tablespace" was faster then the gopher procesing preceeding "drop table". As a result, tablespace file was closed and removed, while "DROP TABLE" still tried to access it.
Solution is to synchronize gopher threads in this case, i.e
all while transaction containing DROP TABLESPACE is  being
processed, all other gophers wait until it completes.
Additionally, a race condition during  CREATE TABLESPACE has
been observed and fixed in this patch. It can happen that
DROP TABLESPACE was completed, but the file is still there
because commit record was not yet processed by a gopher thread.
This resulted into infrequent errors on CREATE with "file already
exists" message.

The solution to this problem is to track number of pending
DROP TABLESPACE commands.During CREATE TABLESPACE, if the file
exists and there are pending DROPs, wait for a short time (10 sec)
in a loop  for file to be deleted.

No testcase is provided with this patch. Had no luck writing
mysql-test-run script reproducing these race conditions.