WL#3726 "DDL locking for all metadata objects".
After review fixes in progress.

Changed open_table() to return bool. This allows more easily to
distinguish cases when this function succeeds but returns no TABLE
instance (in case of view or in case of special kind of open) from
cases when we have an error. Pointer to TABLE instance is now
always returned in TABLE_LIST::table member.

This change allows to get rid of false assumption in open_tables()
implementation and makes it more clear.