Bug#36510 (Stored Procedures: mysql_error_code 0 should be illegal)
This fix is a follow up on Bug 8759.
DECLARE HANDLER can catch exceptions using several syntaxes:
- by SQLSTATE (SQL Standard)
- by error number (MySQL)
Catching succesful completion by error number, aka catching error 0,
is illegal, and is now prevented.

Also, the SQL standard defines the completion condition as a class of
SQLSTATE, so that any state of class '00' and any subclass is a completion
condition. Such completion conditions (for example, '00123') can not be
caught by an exception handler in SQL.