Bug #36968 rpl_temporary_errors.test produces warning in pushbuild
The failure was due to incorrect type of the error reporting function was invoked
at time the slave sql was about to leave.
      
Fixed with correcting that to use the type of logging corrsponding to the error's.

Side effect of this patch is fixes for 4 tests in pb. only recording new results and preventing
execution (on embedded) was needed to be done.