Bug#37191: Failed assertion in CREATE VIEW
It was assumed that if the intersection of all column grants for all 
underlying tables in a view was empty, it implied that the view had 
more privileges on some column than the corresponding underlying table.
Fixed by removing the assumption (which was an assert()) and raising an 
error only when both hold.