Bug#37251 - Livelock between UPDATE and DELETE threads The problem was a bug in deadlock detection code. 2 threads with mutual dependencies doing deadlock at exactly same time both decide there is no deadlock, add themselves to waitFor list, thus building a cycle in it. Both threads will wait for each other until lock wait timeout happens (a deadlock). Meanwhile, any other thread that traverses waitFor list may enter the cycle described above and will loop until deadlock is resolved.
Solution here is to modify deadlock detection to add dependency
first, than check, then remove dependency if cycle is found.
Care is taken to prevent compiler optimizations and hopefully
the combination of volatile specifier and COMPARE_EXHANGE will
generate a memory barrier to prevent reading elements while
they are modified.

Note, that instead of this fancy solution, plain old SyncObject
could be used.  The loop is short, the lock is cheap,
but I can't estimate contention for the moment. If current lock-free
implementation  turns out to be buggy, this alternative can be
considered.

No testcase provided, the situation described in bug report
can not be reliably reproduced.