Temporary work-around for Bug#36403 Compiling Falcon on Solaris fails due to dtrace
To avoid that other people compiling on Solaris have problems with linking
the two DTrace probes found in Falcon we temporarily comment these out
of the code.

Before these should be enabled again there need to be done some work
on how configure and make do linking of DTrace probes for the storage
engines.