bug#36825 - falcon_index_chill_threshold and falcon_record_chill_threshold have values in MB
falcon_index_chill_threshold and falcon_record_chill_threshold accept a value in megabytes. 
This is non-standard with respect to the rest of the server options, prevents from setting 
the variables to lower than 1mb, which prevents testing of the chill/thaw mechanism and 
prevents the user from using the KB, MB and GB modifiers provided by server.
            
falcon_index_chill_threshold and falcon_record_chill_threshold are now changed to accept a value 
in bytes.
      
The tests falcon_options and falcon_options2 are also changed to work accordingly.