BUG#37012 - Online backup: Can't backup to a pipe.
It was impossible to backup to a pipe. Backup code refuse
to overwrite existing files no matter if they're regular files
or named pipes.

This fix allows to write backup image to existing named pipe.

No test case for this fix, as currently there is no good way
in our test suite to determine if OS supports named pipes.