BUG#37918 (Online backup: Incorrect table locking during RESTORE)
Current code uses name locks for locking tables during restore and at the same 
time opens and locks tables for the built-in drivers. It is not correct to 
name-lock some tables and open some other tables at the same time.
      
This patch fixes it by not using name locks, but openning and locking all the