WL4173: Falcon checksum pages
Checksum is updated when one or saveral pages are written to disk.

Checksum is *not* verified  pages are read from disk to 
sector cache,  but first time it is copied from sector cache to 
the page cache.

However preparations are made to verify checksum directly after page 
is read from  from disk, in case sector cache will be made optional
or abandoned.

The algorithm used to calculate checksum  is  home-backed simplified 
fletcher (64bit integers, no 1's complement arithmetic, no care for overflow).
Should be ok for type of errors we expect.