Regression test for issue detected in first implementation of online add/drop index for Falcon, Bug#38039 (Assertion lockState == 0 failed in StorageTableShare::deleteTable SyncObject).
Testing implicit online/offline behavior, i.e. no ONLINE or OFFLINE
keyword is used.

If online add/drop index is disabled, offline will be used instead. 
Hence, the test passes in the current falcon tree, and will pass
if online add/drop index is re-enabled with an appropriate bugfix.

The test will fail if the bug is not fixed when the feature is
re-enabled (ONLINE is implicit when enabled in the Falcon code).