Bug #25998 problems about circle replication Bug #27808 Infinite looping in circular replication
In case of withdrawing one of the servers from the circular multi-master replication group
events generated by the removed server could become unstoppable (bug#25998).
That's because the originator had been the terminator of the own event flow.

Other possibility of the unstoppable event is the cluster replication (bug#27808).
In that case an event generated by a member of a cluster was
replicated to another member, got accepted and executed.
By that same time effects of the event had been already propagated
across the cluster via the cluster communications.
In order to avoid double-applying, a replication event generated 
by a co-member of the cluster should not be accepted.

Both variations of the unstoppable replication event are fixable with 
introducing a new option for CHANGE MASTER: 

IGNORE_SERVER_IDS= (sid_1, sid_2, ... )

The option can be set to the empty list that resets.

Fixed with implementing the feature.

Properties of the feature:

  a. reporting an error if the id of an ignored server is the slave itself and
  its configuration on startup was with --replicate-same-server-id;
  b. overriding the existing IGNORE_SERVER_IDS list by the following 
  CHANGE MASTER ... IGNORE_SERVER_IDS= (list), the empty list arg nullifies
  the current ignored list;
  c. preserving the existing list by CHANGE MASTER w/o IGNORE_SERVER_IDS;
  d. preserving the ignored server ids after RESET SLAVE;
  e. extending SHOW SLAVE STATUS with a new line listing ignored servers;
  f. a new line in master.info with the list of ignored servers;
  g. Differently from --replicate-same-server-id handling, the sql thread is not
  concerned with the ignored server ids, because it's supposed that
  the relay log consists only of events that can not be unstoppable.
  In order to guarantee that, e.g in case of the circular replication with a failing
  server DBA needs to change master necessarily using the new option.
  h. Rotate and FD events originated by the current master listed
  in the ignored list are still relay-logged which does not create
  any termination issue.
  i. The possible list of ignored servers is sorted for the fastest processing of filtering
  algorithm.

Two new lines to show slave status output are added: the list of ignored servers and
the current master server id (yet another feature for the user!).

Use cases for this feature can be found on the bug report page.