Bug#38186 : CREATE TABLESPACE fails sometimes if called directly after DROP TABLESPACE on the same file
The problem is that tablespace files are not deleted directly on client COMMIT.
Deletion is done later, by background(Gopher) thread. Thus the file may still 
exist when "create tablespace" is called.

The solution is to to modify CREATE TABLESPACE and if the file is there and 
there are pending  DROP operation, wait for some seconds until the file is deleted.

An attempt to solve this problem was done previously, but waiting for deletion 
was put on the wrong place. Client thread was waiting for gopher in 
TablespaceManager::dropTablespace, which  can caused  a "soft-deadlock" :
- Client thread was holding syncSysDDL lock and was waiting for Gopher 
  to delete file
- Gopher did not run, was stuck in thread->sleep and waited for Scheduler 
  to wakeup
- Scheduler was doing scavenge and was stuck waiting for syncSysDDL lock 
  in Database:updateCardinalities()

This soft-deadlock was resolved after 10 seconds, client gave up and 
returned with "tablespace file already exists".

The refinement to this solution is given here . Client waits for 
file to be deleted in StorageHandler::createTablespace while no Falcon locks 
are held. Thus the soft-deadlock described is not possible with this solution