Bug#34902: "Backup: crash if view depends on dropped object"
Pre-fix behavior: When adding a view to the backup image fails, 
the object representing the view is deleted but not removed 
from the database catalog. This view object is later tried deleted
as part of catalog delete.
                  
Fix: Add the object and all its dependencies to dependency list
before adding the object to the catalog. If any dependant objects
fail to be added to dep list, the object is not added to the
catalog.