Bug#33053: mysql_upgrade: Does not convert all mysql db tables from utf8mb3 to utf8 Bug#33002: Event Scheduler: Unnecessarily disabled after 5.1 -> 6.0.4 upgrade
The 3-byte Unicode character set formerly known as utf8 is
now known in 6.0 as utf8mb3 and this cause problems if system
tables created in older versions of MySQL are used as the code
that check table compatibility won't accept a different charset
and this was happening for the event table that wasn't being
properly upgraded.

The solution is to ensure that all system tables are upgraded
to use the new utf8 character set rather than the old utf8mb3.
In the future it would be nice if the server could be capable of
handling the character set mishap, but for now we just upgrade
the table as the event subsystem is the only one that checks
if a table is compatible.