Bug#36795 Concurrency issues when starting backups in parallel.
Raise condition on Backup_restore_ctx::mem_alloc since it is static.
The failing backup will set mem_alloc to null when terminating.  The
next time the running backup wants to allocate memory, an assert will
fail.

Makes mem_alloc non-static.  That way, concurrent backups will not
interfere.  This requires that it is possible to bstream_alloc to find
the right Backup_restore_ctx to use.  Fixes that by changing the
static is_running flag to a static pointer, current_op, to the Backup_restore_ctx
of the currently running backup.  If the pointer is null, it means
that no backup is currently running.